Get rid of Lean.HashMap
This is the first step towards eliminating all imports from `Lean`
modules. This is necessary to bring the executable size down to a
reasonable level.

Std.HashMap is coming soon, hopefully we can use that without it bloating the 
executable size.

The next step is to either vendor or replace `ToJson`/`FromJson`